requirements.txt
alipay-sdk-python==3.7.409
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
defusedxml==0.7.1
diff-match-patch==20241021
Django==4.2
django-cors-headers==4.7.0
django-crispy-forms==1.14.0
django-filter==25.1
django-formtools==2.2
django-import-export==4.3.5
django-redis==5.4.0
django-reversion==5.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-extensions==0.7.1
drf-spectacular==0.28.0
future==1.0.0
httplib2==0.9.2
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
mysqlclient==2.2.7
oauthlib==3.2.2
pillow==11.1.0
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.19.1
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
rsa==4.9
sentry-sdk==2.23.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uWSGI==2.0.28
